+++ src/app/dashboard/pages/students/services/student.service.ts
@@ -3,13 +3,14 @@ import { BehaviorSubject, Observable, map, of, take } from 'rxjs';
 import { IStudent } from '../models';
 import { HttpClient } from '@angular/common/http';
 import { environment } from 'src/environments/environments';
+import Swal from 'sweetalert2';
 
 @Injectable({
   providedIn: 'root',
 })
 export class StudentService {
   private _students$ = new BehaviorSubject<IStudent[]>([]);
-  private students$ = this._students$.asObservable();
+  public students$ = this._students$.asObservable();
   constructor(private http: HttpClient) {}
 
   loadStudents(): void {
@@ -24,11 +25,12 @@ export class StudentService {
     return this.students$;
   }
 
-  getStudentById(id: number) {
-    return this.students$.pipe(
-      take(1),
-      map((students) => students.find((u) => u.id === id))
-    );
+  public getStudentOptions(): Observable<IStudent[]> {
+    return this.http.get<IStudent[]>(environment.baseApiUrl + '/students');
+  }
+
+  getStudentById(id: number): Observable<IStudent> {
+    return this.http.get<IStudent>(environment.baseApiUrl + '/students/' + id);
   }
 
   createStudent(payload: IStudent): void {
@@ -50,10 +52,24 @@ export class StudentService {
   }
 
   deleteStudent(id: number): void {
-    this.http.delete(environment.baseApiUrl + '/students/' + id).subscribe({
-      next: (arrayActualizado) => {
-        this.loadStudents();
-      },
+    Swal.fire({
+      title: '¿Estas seguro?',
+      text: 'Se borrará permanentemente',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Si, estoy seguro!',
+      cancelButtonText: 'Cancelar',
+    }).then((result) => {
+      if (result.isConfirmed) {
+        Swal.fire('Eliminado', `Ha sido eliminado con exito`, 'success');
+        this.http.delete(environment.baseApiUrl + '/students/' + id).subscribe({
+          next: (arrayActualizado) => {
+            this.loadStudents();
+          },
+        });
+      }
     });
   }
 }
